+++ src/components/layout/navbar/index.ts
@@ -1,2 +1,2 @@
-export { default as Navigation } from "./navigation";
+export { default as Navigation } from "./Navigation";
 export { default as ConditionalNavigation } from "./ConditionalNavigation";
